refactor 로또 어플리케이션, 입, 출력 부분 구조 변경에 따른 수정
