+++ src/main/java/lotto/README.md
@@ -71,15 +71,23 @@
     - 금액을 입력받는다
     - 로또 장수를 계산한다
     - 로또를 구매한다 
-     
+
+. 상금 Prize
+    - 일치 개수별 상금을 가진다
+    - 일치 개수별 상금을 계산한다 (일치 개수별 상금 * 당첨 수)         
+
 . 당첨 번호 WinningNumbers
     - 당첨 번호를 가지는 컬렉션 객체
     - 문자열 당첨 번호를 숫자 리스트로 변환한다
     - 당첨 번호가 바른 값인지 확인한다.
     - 로또 번호 일치 개수를 확인 한다. getMatchCount();
-        
-. 
 
+. 당첨 결과 WinningResult       
+    - 일치 수(matchCount) 별 당첨 수 (WinningCount)를 가진 컬렉션(Map) 객체
+    - 일치 수 별 당첨 수를 가진 당첨 결과를 만든다. 
+    - 수익율(WinRate)을 계산한다
+          - 수익금(Prize)을 계산한다
+    
         
 ---
 ## 궁금한 점
@@ -90,16 +98,24 @@
 2. Scanner의 상수 선언 관련
     - scanner를 각 메서드 안에 선언 : 쓰고  버림 가능, 중복 코드 발생
     - scanner를 상수로 선언 : 어플리케이션 종료 시점까지 자원 할당 
- 중복되거나 한 클래스 안이더라도 공유된다면 상수 선언이 좋은지?? 
+ 중복되거나 한 클래스 안이더라도 공유된다면 상수 선언이 좋은가요??? 
  
-3. 게터 사용 관련 (쓰면 안되는 이유)
+3. 게터 사용 (쓰면 안되는 이유)
     - 게터 사용을 줄이자! 
     - 게터 : 클래스 변수를 get변수명()의 형태로 리턴하는 메서드
-    - getLottoNumbers() 를 print / showLottoNumbers() 로 사용하는 것은 허용 되는 범위인지?
+    - getLottoNumbers() 를 print / showLottoNumbers() 로 사용하는 것은 허용 되는 범위인지 궁금
      -> 메서드명에 get 대신 다른 키워드를 사용하여 메시지를 전달한다고 봐도 될까요?
-    - 단순 'get' 이라는 키워드를 쓰는건 상관이 없는지~~~~
+    - 단순 'get' 이라는 키워드를 쓰는건 상관이 없는지요~~~~
     
-
-4. LottoMachine
-    - 필요한 상태 값이 없어 서비스로 활용
-    - 인스턴스를 생성하지 않기 위해 유틸 클래스로 사용해야할지..??? 
+    
+4. 정적 팩토리 메서드 관련
+    - 지난 과제(자동차게임) 때 객체를 스스로가 생성하도록 해보라는 리뷰를 받아 리뷰 받은 부분을 리팩토링 진행 했었는데요~
+    - 이번 과제 때도 적용을 해보려고 하다보니 객체로 판단되는 모든 클래스를 다 정적 팩토리 메서드
+        - (생성자 private, 메시지를 갖는 public 정적 메서드 추가) 의 형태로 하면 될지, 
+    범위를 어떻게 하면 좋을지 궁금하여 의견을 듣고 싶습니다.
+     
+    
+5. 올바로 생성되는지 테스트 (지난 리뷰)
+    - 객체 생성 시 올바로 생성됐는지를 테스트하는 것은 
+    생성 후 어떤 런타임 에러가 나는지 여부를 체크하면 되나요??
+    https://github.com/next-step/java-lotto/pull/602#discussion_r434543016
