+++ src/main/java/lotto/ui/InputView.java
@@ -9,12 +9,16 @@ public class InputView {
     public static int enterPurchaseAmount() {
         System.out.println("구입금액을 입력해 주세요.");
         int amount = SCANNER.nextInt();
+        SCANNER.nextLine();
+        /* scanner.nextInt() 사용 후 nextLine()을 사용 시 이 전 입력(nextInt)의 개행 문자가 남아 있는 문제가 있어
+        의미 없는 nextLine() 호출 */
+
         return amount;
     }
 
     public static String enterWinningNumber() {
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
-        String winningNumber = SCANNER.next();
+        String winningNumber = SCANNER.nextLine();
         return winningNumber;
     }
