+++ src/main/java/lotto/ui/ResultView.java
@@ -1,14 +1,12 @@
 package lotto.ui;
 
 import lotto.domain.lotto.LottoTicket;
-import lotto.domain.prize.Prize;
-
-import java.util.Map;
 
 public class ResultView {
-    private ResultView() {}
+    private ResultView() {
+    }
 
-    public static void printLottoQuantity(int quantity){
+    public static void printLottoQuantity(int quantity) {
         System.out.println(quantity + "개를 구매했습니다.");
     }
 
@@ -18,18 +16,19 @@ public class ResultView {
 
     }
 
-    public static void printWinStatistics(Map<Integer, Integer> matchCountMap) {
-        System.out.println("\n당첨 통계\n————");
-        for (Prize prize : Prize.values()) {
-            int matchCount = matchCountMap.getOrDefault(prize.getMatch(), 0);
-            System.out.println(prize.getMatch() + "개 일치(" + prize.getPrice() + "원) - "+matchCount +"개");
+    public static void printWinningRate(double winningRate) {
+        System.out.println("총 수익률은 " + winningRate + "입니다.");
+        if (winningRate < 1) {
+            System.out.println("(기준이 1이기 때문에 결과적으로 손해의 의미임)");
         }
     }
 
-    public static void printWinRate(double winRate) {
-        String result = winRate < 1 ? "손해" : "이득";
-        System.out.println("총 수익률은 " + winRate + "입니다.");
-        System.out.println("(기준이 1이기 때문에 결과적으로 " + result + "의 의미임)");
+    public static void printWinningResult() {
+        System.out.println("\n당첨 통계\n————");
+    }
+
+    public static void printWinningResult(int match, int price, int winningCount) {
+        System.out.println(match + "개 일치(" + price + "원) - " + winningCount + "개");
     }
 
 }
